# backend/app/admin/service/cmd/server/assets/openapi.yaml
...
openapi: 3.0.3
info:
    title: Kratos Admin API
    description: Kratos Admin API
    contact:
        name: tx7do
        url: https://github.com/tx7do/kratos-admin
        email: yanglinbo@gmail.com
    license:
        name: MIT License
        url: https://github.com/tx7do/kratos-admin/blob/master/LICENSE
    version: "1.0"
paths:
...
    /admin/v1/tenants:
        get:
            tags:
                - TenantService
            description: 获取租户列表
            operationId: TenantService_List
            parameters:
                - name: page
                  in: query
                  description: 当前页码
                  schema:
                    type: integer
                    format: int32
                - name: pageSize
                  in: query
                  description: 每一页的行数
                  schema:
                    type: integer
                    format: int32
                - name: query
                  in: query
                  description: AND过滤参数，其语法为json格式的字符串，如：{"key1":"val1","key2":"val2"}，具体请参见：https://github.com/tx7do/go-utils/tree/main/entgo/query/README.md
                  schema:
                    type: string
                - name: or
                  in: query
                  description: OR过滤参数，语法同AND过滤参数。
                  schema:
                    type: string
                - name: orderBy
                  in: query
                  description: 排序条件，其语法为JSON字符串，例如：{"val1", "-val2"}。字段名前加'-'为降序，否则为升序。
                  schema:
                    type: array
                    items:
                        type: string
                - name: noPaging
                  in: query
                  description: 是否不分页，如果为true，则page和pageSize参数无效。
                  schema:
                    type: boolean
                - name: fieldMask
                  in: query
                  description: 字段掩码，其作用为SELECT中的字段，其语法为使用逗号分隔字段名，例如：id,realName,userName。如果为空则选中所有字段，即SELECT *。
                  schema:
                    type: string
                    format: field-mask
                - name: tenantId
                  in: query
                  schema:
                    type: integer
                    format: uint32
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ListTenantResponse'
        post:
            tags:
                - TenantService
            description: 创建租户
            operationId: TenantService_Create
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/CreateTenantRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content: {}
    /admin/v1/tenants/{data.id}:
        put:
            tags:
                - TenantService
            description: 更新租户
            operationId: TenantService_Update
            parameters:
                - name: data.id
                  in: path
                  required: true
                  schema:
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/UpdateTenantRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content: {}
    /admin/v1/tenants/{id}:
        get:
            tags:
                - TenantService
            description: 获取租户数据
            operationId: TenantService_Get
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: integer
                    format: uint32
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Tenant'
        delete:
            tags:
                - TenantService
            description: 删除租户
            operationId: TenantService_Delete
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: integer
                    format: uint32
            responses:
                "200":
                    description: OK
                    content: {}
...
components:
    schemas:
...
        CreateTenantRequest:
            type: object
            properties:
                data:
                    $ref: '#/components/schemas/Tenant'
            description: 创建租户 - 请求
...
        ListTenantResponse:
            type: object
            properties:
                items:
                    type: array
                    items:
                        $ref: '#/components/schemas/Tenant'
                total:
                    type: integer
                    format: uint32
            description: 租户列表 - 答复
...
        Tenant:
            type: object
            properties:
                id:
                    type: integer
                    description: 租户ID
                    format: uint32
                name:
                    type: string
                    description: 租户名称
                code:
                    type: string
                    description: 租户编码
                memberCount:
                    type: integer
                    description: 成员数量
                    format: int32
                status:
                    enum:
                        - ON
                        - OFF
                    type: string
                    default: ON
                    description: 状态
                remark:
                    type: string
                    description: 备注
                createBy:
                    type: integer
                    description: 创建者ID
                    format: uint32
                updateBy:
                    type: integer
                    description: 更新者ID
                    format: uint32
                subscriptionAt:
                    type: string
                    description: 订阅时间
                    format: date-time
                unsubscribeAt:
                    type: string
                    description: 退订时间
                    format: date-time
                createTime:
                    type: string
                    description: 创建时间
                    format: date-time
                updateTime:
                    type: string
                    description: 更新时间
                    format: date-time
                deleteTime:
                    type: string
                    description: 删除时间
                    format: date-time
            description: 租户
...
        UpdateTenantRequest:
            type: object
            properties:
                data:
                    $ref: '#/components/schemas/Tenant'
                updateMask:
                    example: id,realName,userName
                    type: string
                    description: 要更新的字段列表
                    format: field-mask
                allowMissing:
                    type: boolean
                    description: 如果设置为true的时候，资源不存在则会新增(插入)，并且在这种情况下`updateMask`字段将会被忽略。
            description: 更新租户 -请求
...
    securitySchemes:
        OAuth2PasswordBearer:
            type: oauth2
            flows:
                password:
                    tokenUrl: /admin/v1/login
                    refreshUrl: /admin/v1/refresh_token
                    scopes: {}
security:
    - OAuth2PasswordBearer: []
...